refactor: cut  all Dao about Component to dao directory
