+++ src/main/java/org/example/techcare/controller/RepairServlet.java
@@ -7,7 +7,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.example.techcare.model.component.TypeComponent;
-import org.example.techcare.model.component.TypeComponentDAO;
+import org.example.techcare.dao.TypeComponentDAO;
 import org.example.techcare.model.repair.Repair;
 import org.example.techcare.dao.RepairDAO;
